requirements.txt
PyJWT==2.10.1
alembic==1.16.5
fastapi[standard]==0.116.1
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
uvicorn==0.35.0
pydantic-settings==2.10.1